+++ src/main/java/cart/auth/AuthenticationException.java
@@ -1,4 +1,4 @@
-package cart.common.auth;
+package cart.auth;
 
 public class AuthenticationException extends RuntimeException {
