+++ src/main/java/cart/auth/EncodePassword.java
@@ -1,4 +1,4 @@
-package cart.common;
+package cart.auth;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
